Added float to all users input on financa_calculator.py

+++ python_introduction/finance_calculator.py
@@ -1,6 +1,6 @@
 monthly_income = float(input("Enter your monthly income: "))
 Total_monthly_expenses = float(input("Enter your total monthly expenses: "))
 monthly_savings = int(monthly_income - Total_monthly_expenses)
-yearly_savings = int((monthly_savings * 12) + (monthly_savings * 0.05 * 12))
+yearly_savings = ((monthly_savings * 12) + (monthly_savings * 0.05 * 12))
 print (f"your monthly savings are ${monthly_savings}.")
 print (f"Your projected savings after one year, with interest, is: ${yearly_savings}.")
